fix tests, update changelog for release
